Few VS2015 fixes

+++ src/Redists/Constants.cs
@@ -7,7 +7,7 @@
     {
         public const string IntraDelimiter = ":";
         public const string InterDelimiter = "#";
-        public const int BufferSize = 10000;
+        public const long BufferSize = 10000;
         public const string TsPrefix = "ts";
     }
 }
